function parameters

+++ lib/add.go
@@ -8,11 +8,11 @@ func Add(args ...interface{}) (result []interface{}, err error) {
 	if len(args) != 2 {
 		err = errors.New("wrong arguments")
 	}
-	x, xOk := args[0].(float64)
+	x, xOk := args[0].(int64)
 	if !xOk {
 		err = errors.New("wrong parameter")
 	}
-	y, yOk := args[1].(float64)
+	y, yOk := args[1].(int64)
 	if !yOk {
 		err = errors.New("wrong parameter")
 	}
